Version 0.9 is now live!
Overhaul of the connection between SenseGlove_Object and the microcontroller, as well as several bug fixes. Check the changelog for more details!

Changed the Architecture of this Repository slightly.

ChangeLog and Documentation can now be founs in the 'SenseGlove' folder.

You can find previous versions of the SDK in the 'older versions' folder.