v0.10 is now live!
GloveData now has a Unity-friendly wrapper; SenseGlove_Data. We recommend using this wrapper in the future.

Natural limits are added for the fingers and the wrist, through the latter still has issues with gimball lock.

Calibration should no longer crash Untiy now, and more options are available for finger calibration. (However, not all variables are implememnted yet).

Check the changelog for more in-depth changes and bug fixes.

Documentation is updated to be more readable and to reflect the changes in GloveData.

v0.9 can be found in the older versions folder.

Also added .gitIgnore to skip unity .meta files.