v0.15 is now live!
Added a new fancy handmodel, courtesy of the Virtual Dutchmen. In tandem, it sould now be easier to integrate custom hand models into the Sense Glove SDK.

Force feedback is now also integrated in the SDK. It does not require everything to have a rigidbody!

Added a bunch of new Interactables and utility scripts. Updated the SDK Classes image accordingly.

Check the ChangeLog for further details!

Note: When updating to this version of the SDK, it is best to first remove the older version in your project(s), since a few files have changed location!