Version v0.16 is live!
After a long process of development, testing and debugging, the next version of the Sense Glove SDK is finally here, featuring better test scenes, easier integration, haptic- and feedback commands, and a bunch of new experimental 'interactions' to play around with! Check the Changelog for more details.

The scripts folder has been restructured to make it easier for developers to find what they are looking for.