v0.20 is live!
Version 0_20 is now online. This version is compatible with the DK1 Sense Gloves, and no longer compatible with the very first (white) prototypes, which are being pulled from the field.

Added many changes to the internal kinematic models of the Sense Glove, allowing us to calibrate both the hand model and our solvers.

Check the ChangeLog for more details.